Do _not_ disable auto-read until backpressure is working
Produces the issue in `tests/404-hang`. The issue was
that if there is no body reader at all (e.g. because it
is a GET request), auto-read would never get enabled.
But auto-read also affects reading the _next_ request,
even the EOF of the current request.

So this should work differently. Auto-read should stay
on by default, and only when the IncomingMessage hits
it HWM, it should disable it.